services/CreatorCRM/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "dateparser>=1.2.2",
    "email-validator>=2.3.0",
    "flask-uploads>=0.2.1",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "icalendar>=6.3.1",
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-magic>=0.4.27",
    "pytz>=2025.2",
    "redis>=6.4.0",
    "sendgrid>=6.12.4",
    "sqlalchemy>=2.0.43",
    "twilio>=9.7.2",
    "werkzeug>=3.1.3",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "dnspython>=2.7.0",
]
